corrected most warnings, release cleanup
git-svn-id: https://svn.code.sf.net/p/parity/code/trunk@216 26386e8f-9d3d-0410-aaca-840ac0c00a30

+++ parity.binary/CoffImportDirectory.cpp
@@ -66,7 +66,7 @@ namespace parity
 					NativeImportSymbolEntry sym;
 
 					sym.ByOrdinal = (pi->IsOrdinal != 0);
-					sym.Ordinal = (pi->IsOrdinal ? ((unsigned short)(pi->Name & 0xFFFF)) : 0);
+					sym.Ordinal = (pi->IsOrdinal ? (static_cast<unsigned int>(pi->Name & 0xFFFF)) : 0);
 					
 					if(!pi->IsOrdinal) {
 						//
